+++ contracts/factory/examples/schema.rs
@@ -3,8 +3,8 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use wallet_factory::msg::{ExecuteMsg, InstantiateMsg, QueryMsg, WalletListResponse};
-use wallet_factory::state::WalletInfo;
+use vectis_factory::msg::{ExecuteMsg, InstantiateMsg, QueryMsg, WalletListResponse};
+use vectis_factory::state::WalletInfo;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
