+++ contracts/govec/examples/schema.rs
@@ -4,7 +4,7 @@ use std::fs::create_dir_all;
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
 use cw20::{AllAccountsResponse, BalanceResponse, TokenInfoResponse};
-use govec::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use vectis_govec::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
