--- packages/factory/src/lib.rs
@@ -1,15 +1,0 @@
-pub use crate::error::{MigrationMsgError, RelayTxError};
-pub use crate::msg::{
-    CodeIdType, CreateWalletMsg, Guardians, MultiSig, ThresholdAbsoluteCount,
-    WalletFactoryExecuteMsg, WalletFactoryQueryMsg,
-};
-pub use crate::pubkey::pub_key_to_address;
-pub use crate::signature::query_verify_cosmos;
-pub use crate::wallet::{
-    Nonce, ProxyMigrateMsg, ProxyMigrationTxMsg, RelayTransaction, WalletAddr, WalletInfo,
-};
-mod error;
-mod msg;
-mod pubkey;
-mod signature;
-mod wallet;
