+++ packages/vectis/src/error.rs
