+++ packages/vectis/src/lib.rs
@@ -1,0 +1,15 @@
+pub use crate::error::{MigrationMsgError, RelayTxError};
+pub use crate::factory::{
+    CodeIdType, CreateWalletMsg, Guardians, MultiSig, ProxyMigrateMsg, ProxyMigrationTxMsg,
+    ThresholdAbsoluteCount, WalletFactoryExecuteMsg, WalletFactoryQueryMsg,
+};
+pub use crate::govec::StakingOptions;
+pub use crate::pubkey::pub_key_to_address;
+pub use crate::signature::query_verify_cosmos;
+pub use crate::wallet::{Nonce, RelayTransaction, WalletAddr, WalletInfo};
+mod error;
+mod factory;
+mod govec;
+mod pubkey;
+mod signature;
+mod wallet;
