+++ packages/vectis/src/wallet.rs
@@ -1,4 +1,3 @@
-use crate::MigrationMsgError;
 use cosmwasm_std::{Addr, Binary, CanonicalAddr};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -40,28 +39,4 @@ pub struct RelayTransaction {
     /// Serialized signature (message + nonce). Cosmos format (64 bytes).
     /// Cosmos format (secp256k1 verification scheme).
     pub signature: Binary,
-}
-
-#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
-pub enum ProxyMigrationTxMsg {
-    RelayTx(RelayTransaction),
-    DirectMigrationMsg(Binary),
-}
-
-#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
-pub struct ProxyMigrateMsg {
-    pub new_code_id: u64,
-}
-
-impl ProxyMigrateMsg {
-    /// Ensures code id of multisig contract is equal to current factory multisig code id,
-    pub fn ensure_is_supported_proxy_code_id(
-        &self,
-        factory_proxy_code_id: u64,
-    ) -> Result<(), MigrationMsgError> {
-        if factory_proxy_code_id != self.new_code_id {
-            return Err(MigrationMsgError::MismatchProxyCodeId);
-        }
-        Ok(())
-    }
 }
